// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/billing-overview-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface BillingOverviewHttpResponse {
    availableProviders?: Array<string>;
    canStartPremiumTrial?: boolean;
    cancelAtPeriodEnd?: boolean;
    currentPeriodEndUtc?: string | null;
    currentPeriodStartUtc?: string | null;
    isPremium?: boolean;
    manageBillingAvailable?: boolean;
    nextBillingAttemptUtc?: string | null;
    paddleClientToken?: string | null;
    paddleMonthlyPriceId?: string | null;
    paddleYearlyPriceId?: string | null;
    plan?: string | null;
    premiumTrialActive?: boolean;
    premiumTrialEndUtc?: string | null;
    premiumTrialStartUtc?: string | null;
    premiumTrialUsed?: boolean;
    provider?: string;
    renewalEnabled?: boolean;
    subscriptionProvider?: string | null;
    subscriptionStatus?: string | null;
}